requirements.txt
pydantic>=2.8.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
tenacity>=8.2.0
python-dotenv>=1.0.1


